requirements.txt
pinecone-client
python-dotenv
torch
transformers